+++ lineitems.go
@@ -5,7 +5,7 @@ type LineItems struct {
 	ID        string `json:"id,omitempty"`
 	Object    string `json:"object,omitempty"`
 	Name      string `json:"name,omitempty"`
-	UnitPrice int    `json:"unit_price,omitempty"`
+	UnitPrice int64  `json:"unit_price,omitempty"`
 	Quantity  int    `json:"quantity,omitempty"`
 	Deleted   bool   `json:"deleted,omitempty"`
 }
@@ -13,7 +13,7 @@ type LineItems struct {
 //LineItemsParams should be response wrapper for api line_items endpoint
 type LineItemsParams struct {
 	Name      string   `json:"name,omitempty"`
-	UnitPrice int      `json:"unit_price,omitempty"`
+	UnitPrice int64    `json:"unit_price,omitempty"`
 	Quantity  int      `json:"quantity,omitempty"`
 	Metadata  struct{} `json:"metadata,omitempty"`
 }
